+++ test/hemera/error-events.spec.js
@@ -130,10 +130,8 @@ describe('Response error events', function () {
     hemera.ready(() => {
       hemera.on('clientResponseError', function (err) {
         expect(err).to.be.exists()
-        expect(err.name).to.be.equals('FatalError')
-        expect(err.message).to.be.equals('Fatal error')
-        expect(err.cause.name).to.be.equals('Unauthorized')
-        expect(err.cause.message).to.be.equals('test')
+        expect(err.name).to.be.equals('Unauthorized')
+        expect(err.message).to.be.equals('test')
         hemera.close()
         done()
       })
