+++ test/hemera/error-handling.spec.js
@@ -343,40 +343,6 @@ describe('Error handling', function () {
     })
   })
 
-  it('Should crash when an unexpected error thrown during timeout issue', function (done) {
-    const nats = require('nats').connect(authUrl)
-
-    const hemera = new Hemera(nats, {
-      timeout: 20
-    })
-
-    var stub = Sinon.stub(hemera, 'fatal')
-
-    stub.onCall(1)
-
-    stub.returns(true)
-
-    hemera.ready(() => {
-      hemera.act({
-        topic: 'email',
-        cmd: 'send',
-        email: 'foobar@gmail.com',
-        msg: 'Hi!'
-      }, (err, resp) => {
-        expect(err).to.be.exists()
-        // Fatal Error will be throw after the server proceed the msg
-        setTimeout(() => {
-          expect(stub.called).to.be.equals(true)
-          stub.restore()
-          hemera.close()
-          done()
-        }, 500)
-
-        throw (new Error('Test'))
-      })
-    })
-  })
-
   it('Should crash on unhandled business errors', function (done) {
     const nats = require('nats').connect(authUrl)
