+++ test/hemera/error-propagation.spec.js
@@ -1,6 +1,6 @@
 'use strict'
 
-describe.only('Error propagation', function () {
+describe('Error propagation', function () {
   var PORT = 6242
   var flags = ['--user', 'derek', '--pass', 'foobar']
   var authUrl = 'nats://derek:foobar@localhost:' + PORT
